fix: table / order 의존성 역전
