+++ src/main/java/kitchenpos/order/validator/TableValidatorImpl.java
@@ -1,0 +1,46 @@
+package kitchenpos.order.validator;
+
+import java.util.Arrays;
+import java.util.List;
+import java.util.stream.Collectors;
+import kitchenpos.order.domain.OrderStatus;
+import kitchenpos.order.repository.OrderRepository;
+import kitchenpos.table.domain.OrderTable;
+import kitchenpos.table.domain.TableGroup;
+import kitchenpos.table.validator.TableValidator;
+import org.springframework.stereotype.Component;
+
+@Component
+public class TableValidatorImpl implements TableValidator {
+
+    private final OrderRepository orderRepository;
+
+    public TableValidatorImpl(OrderRepository orderRepository) {
+        this.orderRepository = orderRepository;
+    }
+
+    public void validateOrderTableAndOrderStatus(final Long orderTableId) {
+        if (orderRepository.existsByOrderTableIdAndOrderStatusIn(
+            orderTableId, Arrays.asList(OrderStatus.COOKING, OrderStatus.MEAL))) {
+            throw new IllegalArgumentException();
+        }
+    }
+
+
+    public void validateOrderTablesExist(final List<Long> orderTableIds, final List<OrderTable> savedOrderTables) {
+        if (orderTableIds.size() != savedOrderTables.size()) {
+            throw new IllegalArgumentException();
+        }
+    }
+
+    public void validateOrderTableStatus(TableGroup tableGroup) {
+        final List<Long> orderTableIds = tableGroup.getOrderTables().stream()
+            .map(OrderTable::getId)
+            .collect(Collectors.toList());
+
+        if (orderRepository.existsByOrderTableIdInAndOrderStatusIn(
+            orderTableIds, Arrays.asList(OrderStatus.COOKING, OrderStatus.MEAL))) {
+            throw new IllegalArgumentException();
+        }
+    }
+}
