+++ src/main/java/kitchenpos/table/validator/TableValidator.java
@@ -1,45 +1,14 @@
 package kitchenpos.table.validator;
 
-import java.util.Arrays;
 import java.util.List;
-import java.util.stream.Collectors;
-import kitchenpos.order.domain.OrderStatus;
-import kitchenpos.order.repository.OrderRepository;
 import kitchenpos.table.domain.OrderTable;
 import kitchenpos.table.domain.TableGroup;
-import org.springframework.stereotype.Component;
 
-@Component
-public class TableValidator {
+public interface TableValidator {
 
-    private final OrderRepository orderRepository;
+    void validateOrderTableAndOrderStatus(final Long orderTableId);
 
-    public TableValidator(OrderRepository orderRepository) {
-        this.orderRepository = orderRepository;
-    }
+    void validateOrderTablesExist(final List<Long> orderTableIds, final List<OrderTable> savedOrderTables);
 
-    public void validateOrderTableAndOrderStatus(final Long orderTableId) {
-        if (orderRepository.existsByOrderTableIdAndOrderStatusIn(
-            orderTableId, Arrays.asList(OrderStatus.COOKING, OrderStatus.MEAL))) {
-            throw new IllegalArgumentException();
-        }
-    }
-
-
-    public void validateOrderTablesExist(final List<Long> orderTableIds, final List<OrderTable> savedOrderTables) {
-        if (orderTableIds.size() != savedOrderTables.size()) {
-            throw new IllegalArgumentException();
-        }
-    }
-
-    public void validateOrderTableStatus(TableGroup tableGroup) {
-        final List<Long> orderTableIds = tableGroup.getOrderTables().stream()
-            .map(OrderTable::getId)
-            .collect(Collectors.toList());
-
-        if (orderRepository.existsByOrderTableIdInAndOrderStatusIn(
-            orderTableIds, Arrays.asList(OrderStatus.COOKING, OrderStatus.MEAL))) {
-            throw new IllegalArgumentException();
-        }
-    }
+    void validateOrderTableStatus(TableGroup tableGroup);
 }
